Updating flow defintions
Pushed from Domino: https://lsdemo.domino-eval.com/workspace/ross_domino/SCE_Demo_Refactor?executionId=689b7dd7418b56404efcb901
