+++ flows/flow_1_dev.py
@@ -10,22 +10,20 @@ from flytekitplugins.domino.artifact import Artifact, DATA, MODEL, REPORT
 environment_name="SAS Analytics Pro"
 hardware_tier_name="Small"
 
-
 # Enter the command below to run this Flow. There is a single Flow input parameter for the SDTM Dataset snapshot
-# pyflyte run --remote ./flows/flow_3.py ADaM_only_QC --sdtm_dataset_snapshot /mnt/imported/data/SDTMBLIND
+# pyflyte run --remote ./flows/flow_1_dev.py ADaM_only --sdtm_dataset_snapshot /mnt/imported/data/SDTMBLIND 
 
 # If you want to give the run a name, then use this command and replace the MY_CUSTOM_NAME argument
-# pyflyte run --remote --name MY_CUSTOM_NAME ./flows/flow_3.py ADaM_only_QC --sdtm_dataset_snapshot /mnt/imported/data/SDTMBLIND
+# pyflyte run --remote --name ENTER_RUN_NAME ./flows/flow_1_dev.py ADaM_only --sdtm_dataset_snapshot /mnt/imported/data/SDTMBLIND 
 
 
-# Define two Flow Artifacts called ADaM Dataset and QC ADaM Dataset to tag and group ADaM outputs respectively
+# Define one Flow Artifact called ADaM Dataset to tag and group all of the ADAM task outputs as
 DataArtifact = Artifact("ADaM Datasets", DATA)
-QCDataArtifact = Artifact("QC ADaM Datasets", DATA)
+
 
 @workflow
-def ADaM_only_QC(sdtm_dataset_snapshot: str):
+def ADaM_only(sdtm_dataset_snapshot: str):
 
-    #PROD 
     adsl_task = run_domino_job_task(
         flyte_task_name="Create ADSL Dataset",
         command="prod/adam/ADSL.sas",
@@ -34,9 +32,8 @@ def ADaM_only_QC(sdtm_dataset_snapshot: str):
         hardware_tier_name=hardware_tier_name,
         environment_name=environment_name,
         use_project_defaults_for_omitted=True
-    ) 
-
-    #PROD 
+    )
+ 
     adae_task = run_domino_job_task(
         flyte_task_name="Create ADAE Dataset",
         command="prod/adam/ADAE.sas",
@@ -47,7 +44,7 @@ def ADaM_only_QC(sdtm_dataset_snapshot: str):
         environment_name=environment_name,
         use_project_defaults_for_omitted=True
     )
-    #PROD 
+    
     adcm_task = run_domino_job_task(
         flyte_task_name="Create ADCM Dataset",
         command="prod/adam/ADCM.sas",
@@ -58,7 +55,7 @@ def ADaM_only_QC(sdtm_dataset_snapshot: str):
         environment_name=environment_name,
         use_project_defaults_for_omitted=True
     )
-    #PROD 
+
     adlb_task = run_domino_job_task(
         flyte_task_name="Create ADLB Dataset",
         command="prod/adam/ADLB.sas",
@@ -67,9 +64,9 @@ def ADaM_only_QC(sdtm_dataset_snapshot: str):
         output_specs=[Output(name="adlb_dataset", type=DataArtifact.File(name="adlb", type="sas7bdat"))],
         hardware_tier_name=hardware_tier_name,
         environment_name=environment_name,
-        use_project_defaults_for_omitted=True,
+        use_project_defaults_for_omitted=True
     )
-    #PROD 
+
     admh_task = run_domino_job_task(
         flyte_task_name="Create ADMH Dataset",
         command="prod/adam/ADMH.sas",
@@ -80,7 +77,7 @@ def ADaM_only_QC(sdtm_dataset_snapshot: str):
         environment_name=environment_name,
         use_project_defaults_for_omitted=True
     )
-    #PROD 
+
     advs_task = run_domino_job_task(
         flyte_task_name="Create ADVS Dataset",
         command="prod/adam/ADVS.sas",
@@ -91,71 +88,5 @@ def ADaM_only_QC(sdtm_dataset_snapshot: str):
         environment_name=environment_name,
         use_project_defaults_for_omitted=True
     )
-    #QC 
-    qc_adsl_task = run_domino_job_task(
-        flyte_task_name="Create QC ADSL Dataset",
-        command="qc/adam/qc_ADSL.sas",
-        inputs=[Input(name="sdtm_snapshot_task_input", type=str, value=sdtm_dataset_snapshot)],
-        output_specs=[Output(name="qc_adsl_dataset", type=QCDataArtifact.File(name="qc_adsl", type="sas7bdat"))],
-        hardware_tier_name=hardware_tier_name,
-        environment_name=environment_name,
-        use_project_defaults_for_omitted=True,
-    ) 
- 
-    #QC 
-    qc_adae_task = run_domino_job_task(
-        flyte_task_name="Create QC ADAE Dataset",
-        command="qc/adam/qc_ADAE.sas",
-        inputs=[Input(name="sdtm_snapshot_task_input", type=str, value=sdtm_dataset_snapshot),
-                Input(name="qc_adsl_dataset", type=FlyteFile[TypeVar("sas7bdat")], value=qc_adsl_task["qc_adsl_dataset"])],
-        output_specs=[Output(name="qc_adae_dataset", type=QCDataArtifact.File(name="qc_adae", type="sas7bdat"))],
-        hardware_tier_name=hardware_tier_name,
-        environment_name=environment_name,
-        use_project_defaults_for_omitted=True
-    )
-    #QC 
-    qc_adcm_task = run_domino_job_task(
-        flyte_task_name="Create QC ADCM Dataset",
-        command="qc/adam/qc_ADCM.sas",
-        inputs=[Input(name="sdtm_snapshot_task_input", type=str, value=sdtm_dataset_snapshot),
-                Input(name="qc_adsl_dataset", type=FlyteFile[TypeVar("sas7bdat")], value=qc_adsl_task["qc_adsl_dataset"])],
-        output_specs=[Output(name="qc_adcm_dataset", type=QCDataArtifact.File(name="qc_adcm", type="sas7bdat"))],
-        hardware_tier_name=hardware_tier_name,
-        environment_name=environment_name,
-        use_project_defaults_for_omitted=True
-    )
-    #QC 
-    qc_adlb_task = run_domino_job_task(
-        flyte_task_name="Create QC ADLB Dataset",
-        command="qc/adam/qc_ADLB.sas",
-        inputs=[Input(name="sdtm_snapshot_task_input", type=str, value=sdtm_dataset_snapshot),
-                Input(name="qc_adsl_dataset", type=FlyteFile[TypeVar("sas7bdat")], value=qc_adsl_task["qc_adsl_dataset"])],
-        output_specs=[Output(name="qc_adlb_dataset", type=QCDataArtifact.File(name="qc_adlb", type="sas7bdat"))],
-        hardware_tier_name=hardware_tier_name,
-        environment_name=environment_name,
-        use_project_defaults_for_omitted=True
-    )
-    #QC 
-    qc_admh_task = run_domino_job_task(
-        flyte_task_name="Create QC ADMH Dataset",
-        command="qc/adam/qc_ADMH.sas",
-        inputs=[Input(name="sdtm_snapshot_task_input", type=str, value=sdtm_dataset_snapshot),
-                Input(name="qc_adsl_dataset", type=FlyteFile[TypeVar("sas7bdat")], value=qc_adsl_task["qc_adsl_dataset"])],
-        output_specs=[Output(name="qc_admh_dataset", type=QCDataArtifact.File(name="qc_admh", type="sas7bdat"))],
-        hardware_tier_name=hardware_tier_name,
-        environment_name=environment_name,
-        use_project_defaults_for_omitted=True
-    )
-    #QC 
-    qc_advs_task = run_domino_job_task(
-        flyte_task_name="Create QC ADVS Dataset",
-        command="qc/adam/qc_ADVS.sas",
-        inputs=[Input(name="sdtm_snapshot_task_input", type=str, value=sdtm_dataset_snapshot),
-                Input(name="qc_adsl_dataset", type=FlyteFile[TypeVar("sas7bdat")], value=qc_adsl_task["qc_adsl_dataset"])],
-        output_specs=[Output(name="qc_advs_dataset", type=QCDataArtifact.File(name="qc_advs", type="sas7bdat"))],
-        hardware_tier_name=hardware_tier_name,
-        environment_name=environment_name,
-        use_project_defaults_for_omitted=True
-    )
 
-    return 
+    return
