+++ flows/flow_2_dev.py
@@ -1,6 +1,6 @@
 from flytekit import workflow
 from flytekit.types.file import FlyteFile
-from typing import TypeVar, NamedTuple, Tuple
+from typing import TypeVar, NamedTuple
 from flytekitplugins.domino.helpers import Input, Output, run_domino_job_task
 from flytekitplugins.domino.task import DominoJobConfig, DominoJobTask, GitRef, EnvironmentRevisionSpecification, EnvironmentRevisionType, DatasetSnapshot
 from flytekitplugins.domino.artifact import Artifact, DATA, MODEL, REPORT
@@ -10,19 +10,20 @@ from flytekitplugins.domino.artifact import Artifact, DATA, MODEL, REPORT
 environment_name="SAS Analytics Pro"
 hardware_tier_name="Small"
 
-# Enter the command below to run this Flow. There is a single Flow input parameter for the SDTM Dataset snapshot
-# pyflyte run --remote ./flows/flow_1.py ADaM_only --sdtm_dataset_snapshot /mnt/imported/data/SDTMBLIND 
+
+# Enter the command below to run this Flow. There are two Flow input parameters. One for the SDTM Dataset snapshot and one for the METADATA dataset snapshot.
+# pyflyte run --remote ./flows/flow_2_dev.py ADaM_TFL --sdtm_dataset_snapshot /mnt/imported/data/SDTMBLIND --metadata_snapshot /mnt/data/METADATA 
 
 # If you want to give the run a name, then use this command and replace the MY_CUSTOM_NAME argument
-# pyflyte run --remote --name MY_CUSTOM_NAME ./flows/flow_1.py ADaM_only --sdtm_dataset_snapshot /mnt/imported/data/SDTMBLIND 
+# pyflyte run --remote --name ENTER_RUN_NAME ./flows/flow_2_dev.py ADaM_TFL --sdtm_dataset_snapshot /mnt/imported/data/SDTMBLIND --metadata_snapshot /mnt/data/METADATA 
 
 
-# Define one Flow Artifact called ADaM Dataset to tag and group all of the ADAM task outputs as
+# Define two Flow Artifacts called ADaM Dataset and TFL Reports to tag and group ADaM and TFL outputs respectively
 DataArtifact = Artifact("ADaM Datasets", DATA)
-
+ReportArtifact = Artifact("TFL Reports", REPORT)
 
 @workflow
-def ADaM_only(sdtm_dataset_snapshot: str):
+def ADaM_TFL(sdtm_dataset_snapshot: str, metadata_snapshot: str): 
 
     adsl_task = run_domino_job_task(
         flyte_task_name="Create ADSL Dataset",
@@ -32,8 +33,8 @@ def ADaM_only(sdtm_dataset_snapshot: str):
         hardware_tier_name=hardware_tier_name,
         environment_name=environment_name,
         use_project_defaults_for_omitted=True
-    )
- 
+    ) 
+
     adae_task = run_domino_job_task(
         flyte_task_name="Create ADAE Dataset",
         command="prod/adam/ADAE.sas",
@@ -84,6 +85,40 @@ def ADaM_only(sdtm_dataset_snapshot: str):
         inputs=[Input(name="sdtm_snapshot_task_input", type=str, value=sdtm_dataset_snapshot),
                 Input(name="adsl_dataset", type=FlyteFile[TypeVar("sas7bdat")], value=adsl_task["adsl_dataset"])],
         output_specs=[Output(name="advs_dataset", type=DataArtifact.File(name="advs", type="sas7bdat"))],
+        hardware_tier_name=hardware_tier_name,
+        environment_name=environment_name,
+        use_project_defaults_for_omitted=True
+    )
+
+    t_pop_task = run_domino_job_task(
+        flyte_task_name="Create T_POP Report",
+        command="prod/tfl/t_pop.sas",
+        inputs=[Input(name="adsl_dataset", type=FlyteFile[TypeVar("sas7bdat")], value=adsl_task["adsl_dataset"]),
+                Input(name="metadata_snapshot", type=str, value=metadata_snapshot)],
+        output_specs=[Output(name="t_pop", type=ReportArtifact.File(name="t_pop", type="pdf"))],
+        hardware_tier_name=hardware_tier_name,
+        environment_name=environment_name,
+        use_project_defaults_for_omitted=True
+    )
+
+    t_ae_rel_task = run_domino_job_task(
+        flyte_task_name="Create T_AE_REL Report",
+        command="prod/tfl/t_ae_rel.sas",
+        inputs=[Input(name="adsl_dataset", type=FlyteFile[TypeVar("sas7bdat")], value=adsl_task["adsl_dataset"]),
+                Input(name="adae_dataset", type=FlyteFile[TypeVar("sas7bdat")], value=adae_task["adae_dataset"]),
+                Input(name="metadata_snapshot", type=str, value=metadata_snapshot)],
+        output_specs=[Output(name="t_ae_rel", type=ReportArtifact.File(name="t_ae_rel", type="pdf"))],
+        hardware_tier_name=hardware_tier_name,
+        environment_name=environment_name,
+        use_project_defaults_for_omitted=True
+    )
+
+    t_vscat_task = run_domino_job_task(
+        flyte_task_name="Create T_VSCAT Report",
+        command="prod/tfl/t_vscat.sas",
+        inputs=[Input(name="advs_dataset", type=FlyteFile[TypeVar("sas7bdat")], value=advs_task["advs_dataset"]),
+                Input(name="metadata_snapshot", type=str, value=metadata_snapshot)],
+        output_specs=[Output(name="t_vscat", type=ReportArtifact.File(name="t_vscat", type="pdf"))],
         hardware_tier_name=hardware_tier_name,
         environment_name=environment_name,
         use_project_defaults_for_omitted=True
