+++ flows/flow_3_dev.py
@@ -1,6 +1,6 @@
 from flytekit import workflow
 from flytekit.types.file import FlyteFile
-from typing import TypeVar, NamedTuple
+from typing import TypeVar, NamedTuple, Tuple
 from flytekitplugins.domino.helpers import Input, Output, run_domino_job_task
 from flytekitplugins.domino.task import DominoJobConfig, DominoJobTask, GitRef, EnvironmentRevisionSpecification, EnvironmentRevisionType, DatasetSnapshot
 from flytekitplugins.domino.artifact import Artifact, DATA, MODEL, REPORT
@@ -11,20 +11,21 @@ environment_name="SAS Analytics Pro"
 hardware_tier_name="Small"
 
 
-# Enter the command below to run this Flow. There are two Flow input parameters. One for the SDTM Dataset snapshot and one for the METADATA dataset snapshot.
-# pyflyte run --remote ./flows/flow_2.py ADaM_TFL --sdtm_dataset_snapshot /mnt/imported/data/SDTMBLIND --metadata_snapshot /mnt/data/METADATA 
+# Enter the command below to run this Flow. There is a single Flow input parameter for the SDTM Dataset snapshot
+# pyflyte run --remote ./flows/flow_3_dev.py ADaM_only_QC --sdtm_dataset_snapshot /mnt/imported/data/SDTMBLIND
 
 # If you want to give the run a name, then use this command and replace the MY_CUSTOM_NAME argument
-# pyflyte run --remote --name MY_CUSTOM_NAME ./flows/flow_2.py ADaM_TFL --sdtm_dataset_snapshot /mnt/imported/data/SDTMBLIND --metadata_snapshot /mnt/data/METADATA 
+# pyflyte run --remote --name ENTER_RUN_NAME ./flows/flow_3_dev.py ADaM_only_QC --sdtm_dataset_snapshot /mnt/imported/data/SDTMBLIND
 
 
-# Define two Flow Artifacts called ADaM Dataset and TFL Reports to tag and group ADaM and TFL outputs respectively
+# Define two Flow Artifacts called ADaM Dataset and QC ADaM Dataset to tag and group ADaM outputs respectively
 DataArtifact = Artifact("ADaM Datasets", DATA)
-ReportArtifact = Artifact("TFL Reports", REPORT)
+QCDataArtifact = Artifact("QC ADaM Datasets", DATA)
 
 @workflow
-def ADaM_TFL(sdtm_dataset_snapshot: str, metadata_snapshot: str): 
+def ADaM_only_QC(sdtm_dataset_snapshot: str):
 
+    #PROD 
     adsl_task = run_domino_job_task(
         flyte_task_name="Create ADSL Dataset",
         command="prod/adam/ADSL.sas",
@@ -35,6 +36,7 @@ def ADaM_TFL(sdtm_dataset_snapshot: str, metadata_snapshot: str):
         use_project_defaults_for_omitted=True
     ) 
 
+    #PROD 
     adae_task = run_domino_job_task(
         flyte_task_name="Create ADAE Dataset",
         command="prod/adam/ADAE.sas",
@@ -45,7 +47,7 @@ def ADaM_TFL(sdtm_dataset_snapshot: str, metadata_snapshot: str):
         environment_name=environment_name,
         use_project_defaults_for_omitted=True
     )
-    
+    #PROD 
     adcm_task = run_domino_job_task(
         flyte_task_name="Create ADCM Dataset",
         command="prod/adam/ADCM.sas",
@@ -56,7 +58,7 @@ def ADaM_TFL(sdtm_dataset_snapshot: str, metadata_snapshot: str):
         environment_name=environment_name,
         use_project_defaults_for_omitted=True
     )
-
+    #PROD 
     adlb_task = run_domino_job_task(
         flyte_task_name="Create ADLB Dataset",
         command="prod/adam/ADLB.sas",
@@ -65,9 +67,9 @@ def ADaM_TFL(sdtm_dataset_snapshot: str, metadata_snapshot: str):
         output_specs=[Output(name="adlb_dataset", type=DataArtifact.File(name="adlb", type="sas7bdat"))],
         hardware_tier_name=hardware_tier_name,
         environment_name=environment_name,
-        use_project_defaults_for_omitted=True
+        use_project_defaults_for_omitted=True,
     )
-
+    #PROD 
     admh_task = run_domino_job_task(
         flyte_task_name="Create ADMH Dataset",
         command="prod/adam/ADMH.sas",
@@ -78,7 +80,7 @@ def ADaM_TFL(sdtm_dataset_snapshot: str, metadata_snapshot: str):
         environment_name=environment_name,
         use_project_defaults_for_omitted=True
     )
-
+    #PROD 
     advs_task = run_domino_job_task(
         flyte_task_name="Create ADVS Dataset",
         command="prod/adam/ADVS.sas",
@@ -89,39 +91,71 @@ def ADaM_TFL(sdtm_dataset_snapshot: str, metadata_snapshot: str):
         environment_name=environment_name,
         use_project_defaults_for_omitted=True
     )
-
-    t_pop_task = run_domino_job_task(
-        flyte_task_name="Create T_POP Report",
-        command="prod/tfl/t_pop.sas",
-        inputs=[Input(name="adsl_dataset", type=FlyteFile[TypeVar("sas7bdat")], value=adsl_task["adsl_dataset"]),
-                Input(name="metadata_snapshot", type=str, value=metadata_snapshot)],
-        output_specs=[Output(name="t_pop", type=ReportArtifact.File(name="t_pop", type="pdf"))],
+    #QC 
+    qc_adsl_task = run_domino_job_task(
+        flyte_task_name="Create QC ADSL Dataset",
+        command="qc/adam/qc_ADSL.sas",
+        inputs=[Input(name="sdtm_snapshot_task_input", type=str, value=sdtm_dataset_snapshot)],
+        output_specs=[Output(name="qc_adsl_dataset", type=QCDataArtifact.File(name="qc_adsl", type="sas7bdat"))],
+        hardware_tier_name=hardware_tier_name,
+        environment_name=environment_name,
+        use_project_defaults_for_omitted=True,
+    ) 
+ 
+    #QC 
+    qc_adae_task = run_domino_job_task(
+        flyte_task_name="Create QC ADAE Dataset",
+        command="qc/adam/qc_ADAE.sas",
+        inputs=[Input(name="sdtm_snapshot_task_input", type=str, value=sdtm_dataset_snapshot),
+                Input(name="qc_adsl_dataset", type=FlyteFile[TypeVar("sas7bdat")], value=qc_adsl_task["qc_adsl_dataset"])],
+        output_specs=[Output(name="qc_adae_dataset", type=QCDataArtifact.File(name="qc_adae", type="sas7bdat"))],
+        hardware_tier_name=hardware_tier_name,
+        environment_name=environment_name,
+        use_project_defaults_for_omitted=True
+    )
+    #QC 
+    qc_adcm_task = run_domino_job_task(
+        flyte_task_name="Create QC ADCM Dataset",
+        command="qc/adam/qc_ADCM.sas",
+        inputs=[Input(name="sdtm_snapshot_task_input", type=str, value=sdtm_dataset_snapshot),
+                Input(name="qc_adsl_dataset", type=FlyteFile[TypeVar("sas7bdat")], value=qc_adsl_task["qc_adsl_dataset"])],
+        output_specs=[Output(name="qc_adcm_dataset", type=QCDataArtifact.File(name="qc_adcm", type="sas7bdat"))],
+        hardware_tier_name=hardware_tier_name,
+        environment_name=environment_name,
+        use_project_defaults_for_omitted=True
+    )
+    #QC 
+    qc_adlb_task = run_domino_job_task(
+        flyte_task_name="Create QC ADLB Dataset",
+        command="qc/adam/qc_ADLB.sas",
+        inputs=[Input(name="sdtm_snapshot_task_input", type=str, value=sdtm_dataset_snapshot),
+                Input(name="qc_adsl_dataset", type=FlyteFile[TypeVar("sas7bdat")], value=qc_adsl_task["qc_adsl_dataset"])],
+        output_specs=[Output(name="qc_adlb_dataset", type=QCDataArtifact.File(name="qc_adlb", type="sas7bdat"))],
+        hardware_tier_name=hardware_tier_name,
+        environment_name=environment_name,
+        use_project_defaults_for_omitted=True
+    )
+    #QC 
+    qc_admh_task = run_domino_job_task(
+        flyte_task_name="Create QC ADMH Dataset",
+        command="qc/adam/qc_ADMH.sas",
+        inputs=[Input(name="sdtm_snapshot_task_input", type=str, value=sdtm_dataset_snapshot),
+                Input(name="qc_adsl_dataset", type=FlyteFile[TypeVar("sas7bdat")], value=qc_adsl_task["qc_adsl_dataset"])],
+        output_specs=[Output(name="qc_admh_dataset", type=QCDataArtifact.File(name="qc_admh", type="sas7bdat"))],
+        hardware_tier_name=hardware_tier_name,
+        environment_name=environment_name,
+        use_project_defaults_for_omitted=True
+    )
+    #QC 
+    qc_advs_task = run_domino_job_task(
+        flyte_task_name="Create QC ADVS Dataset",
+        command="qc/adam/qc_ADVS.sas",
+        inputs=[Input(name="sdtm_snapshot_task_input", type=str, value=sdtm_dataset_snapshot),
+                Input(name="qc_adsl_dataset", type=FlyteFile[TypeVar("sas7bdat")], value=qc_adsl_task["qc_adsl_dataset"])],
+        output_specs=[Output(name="qc_advs_dataset", type=QCDataArtifact.File(name="qc_advs", type="sas7bdat"))],
         hardware_tier_name=hardware_tier_name,
         environment_name=environment_name,
         use_project_defaults_for_omitted=True
     )
 
-    t_ae_rel_task = run_domino_job_task(
-        flyte_task_name="Create T_AE_REL Report",
-        command="prod/tfl/t_ae_rel.sas",
-        inputs=[Input(name="adsl_dataset", type=FlyteFile[TypeVar("sas7bdat")], value=adsl_task["adsl_dataset"]),
-                Input(name="adae_dataset", type=FlyteFile[TypeVar("sas7bdat")], value=adae_task["adae_dataset"]),
-                Input(name="metadata_snapshot", type=str, value=metadata_snapshot)],
-        output_specs=[Output(name="t_ae_rel", type=ReportArtifact.File(name="t_ae_rel", type="pdf"))],
-        hardware_tier_name=hardware_tier_name,
-        environment_name=environment_name,
-        use_project_defaults_for_omitted=True
-    )
-
-    t_vscat_task = run_domino_job_task(
-        flyte_task_name="Create T_VSCAT Report",
-        command="prod/tfl/t_vscat.sas",
-        inputs=[Input(name="advs_dataset", type=FlyteFile[TypeVar("sas7bdat")], value=advs_task["advs_dataset"]),
-                Input(name="metadata_snapshot", type=str, value=metadata_snapshot)],
-        output_specs=[Output(name="t_vscat", type=ReportArtifact.File(name="t_vscat", type="pdf"))],
-        hardware_tier_name=hardware_tier_name,
-        environment_name=environment_name,
-        use_project_defaults_for_omitted=True
-    )
-
-    return
+    return
